Implement ConfigureMgr for container and merge it with native
