--- internal/controller/configuremgr/container/container.go
@@ -1,45 +1,0 @@
-/*******************************************************************************
- * Copyright 2019 Samsung Electronics All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- *******************************************************************************/
-
-// Package container provides container specific functions for configuremgr
-package container
-
-import (
-	"github.com/lf-edge/edge-home-orchestration-go/internal/controller/configuremgr"
-)
-
-// ConfigureMgr has config folder path
-type ConfigureMgr struct {
-	confpath string
-}
-
-var configuremgrObj *ConfigureMgr
-
-func init() {
-	configuremgrObj = new(ConfigureMgr)
-}
-
-// GetInstance set configpath and gives ConfigureMgrs Singletone instance
-func GetInstance(configPath string) *ConfigureMgr {
-	configuremgrObj.confpath = configPath
-	return configuremgrObj
-}
-
-// Watch implements Watcher interface with ConfigureMgr struct
-func (cfgMgr ConfigureMgr) Watch(notifier configuremgr.Notifier) {
-	// Do nothing because it is Watcher interface
-}
