+++ internal/controller/configuremgr/testdata/mysum/libmysum.so.1.0.1
