+++ internal/controller/configuremgr/testdata/mysum/mysum.conf
@@ -4,6 +4,7 @@ ConfVersion=v0.0                    ; Version of Configuration file
 
 [ServiceInfo]
 ServiceName=HelloWorldService       ; Name of distributed service
+ExecType=native                     ; Execution type of service
 
 [ScoringMethod]
 LibFile=./libmysum.so               ; Library file name
