+++ test/container/datastorage/datastorage.conf
@@ -3,10 +3,8 @@
 ConfVersion=v0.0                    ; Version of Configuration file
 
 [ServiceInfo]
-ServiceName=ls                      ; Name of distributed service
-AllowedRequester=bash,curl
-ExecutableFileName=ls
-ExecType=native
+ServiceName=DataStorage             ; Name of distributed service
+ExecType=container
 
 [ScoringMethod]
 LibFile=                            ; Library file name
